requirements.txt
python-telegram-bot
loguru
pydantic
pydantic-settings
aiosmtplib